TASK-2: Add status columns to allow completing a todo. Show only the todos for the current user

+++ alayatodo/views.py
@@ -45,7 +45,7 @@ def logout():
 
 @app.route('/todo/<id>', methods=['GET'])
 def todo(id):
-    cur = g.db.execute("SELECT * FROM todos WHERE id ='%s'" % id)
+    cur = g.db.execute("SELECT * FROM todos WHERE id ='%s' AND user_id='%s'" % (id, session['user']['id']))
     todo = cur.fetchone()
     return render_template('todo.html', todo=todo)
 
@@ -55,7 +55,7 @@ def todo(id):
 def todos():
     if not session.get('logged_in'):
         return redirect('/login')
-    cur = g.db.execute("SELECT * FROM todos")
+    cur = g.db.execute("SELECT * FROM todos WHERE user_id = '%s'" % session['user']['id'])
     todos = cur.fetchall()
     return render_template('todos.html', todos=todos)
 
@@ -79,6 +79,16 @@ def todos_POST():
 def todo_delete(id):
     if not session.get('logged_in'):
         return redirect('/login')
-    g.db.execute("DELETE FROM todos WHERE id ='%s'" % id)
+    g.db.execute("DELETE FROM todos WHERE id ='%s' AND user_id='%s'" % (id, session['user']['id']))
+    g.db.commit()
+    return redirect('/todo')
+
+
+@app.route('/todo/complete/<id>', methods=['GET'])
+def todo_complete(id):
+    if not session.get('logged_in'):
+        return redirect('/login')
+
+    g.db.execute("UPDATE todos SET status = %d WHERE id = '%s' AND user_id = '%s'" % (1, id, session['user']['id']))
     g.db.commit()
     return redirect('/todo')
